backend/requirements.txt
fastapi==0.109.2
python-multipart==0.0.9
uvicorn==0.27.1
google-generativeai==0.3.2
Pillow==10.2.0
pytesseract==0.3.10
PyMuPDF==1.23.26
faiss-cpu==1.7.4
numpy==1.26.4
python-dotenv==1.0.1